+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ### v1.3.4
 #### Added
 - Typedoc for generating code documentation.
+#### Updated
+- Updated the Panel components to take in new props and used the Header from Reusable Components.
+- Updated the PlsIDForm to use the `uuid` prop as an ID to its `Input` component for accessibility purposes.
 
 ### v1.3.3
 #### Added